cleanup test view
