+++ Example/Example/Logs/LogsView.swift
@@ -15,16 +15,16 @@ struct LogsView: View {
     @State private var expandedIds: Set<UUID> = []
     @State private var isShowingShareSheet = false
     @State private var didJustCopy = false
-    
+
     enum Filter: String, CaseIterable, Identifiable {
         case all = "All"
         case actions = "Actions"
         case events = "Events"
         case locations = "Locations"
         case logs = "Logs"
-        
+
         var id: String { rawValue }
-        
+
         func includes(_ kind: ConsoleEntry.Kind) -> Bool {
             switch self {
             case .all: return true
@@ -35,7 +35,7 @@ struct LogsView: View {
             }
         }
     }
-    
+
     var body: some View {
         VStack(spacing: 0) {
             header
@@ -47,9 +47,9 @@ struct LogsView: View {
             ActivityShareSheet(activityItems: [ConsoleEntry.formatForExport(filtered)])
         }
     }
-    
+
     // MARK: - Subviews
-    
+
     private var header: some View {
         HStack(spacing: 12) {
             Text("Console").font(.title2.weight(.semibold))
@@ -87,18 +87,6 @@ struct LogsView: View {
         .padding(.bottom, 8)
     }
 
-    private func copyToPasteboard() {
-        UIPasteboard.general.string = ConsoleEntry.formatForExport(filtered)
-        withAnimation {
-            didJustCopy = true
-        }
-        DispatchQueue.main.asyncAfter(deadline: .now() + 1.5) {
-            withAnimation {
-                didJustCopy = false
-            }
-        }
-    }
-    
     private var filterRow: some View {
         ScrollView(.horizontal, showsIndicators: false) {
             HStack(spacing: 8) {
@@ -119,7 +107,7 @@ struct LogsView: View {
             .padding(.bottom, 8)
         }
     }
-    
+
     @ViewBuilder
     private var content: some View {
         if filtered.isEmpty {
@@ -127,13 +115,13 @@ struct LogsView: View {
         } else {
             List {
                 ForEach(filtered) { entry in
-                    entryRow(entry)
+                    ConsoleEntryRow(entry: entry, isExpanded: expansionBinding(for: entry))
                 }
             }
             .listStyle(.plain)
         }
     }
-    
+
     private var emptyState: some View {
         VStack(spacing: 12) {
             Image(systemName: "tray")
@@ -150,78 +138,36 @@ struct LogsView: View {
         }
         .frame(maxWidth: .infinity, maxHeight: .infinity)
     }
-    
-    private func entryRow(_ entry: ConsoleEntry) -> some View {
-        let isExpanded = expandedIds.contains(entry.id)
-        let canExpand = entry.detail != nil
-        
-        return VStack(alignment: .leading, spacing: 6) {
-            HStack(alignment: .top, spacing: 8) {
-                Image(systemName: entry.kind.iconName)
-                    .foregroundColor(entry.kind.tintColor)
-                    .frame(width: 18)
-                VStack(alignment: .leading, spacing: 2) {
-                    HStack(spacing: 6) {
-                        Text(timeString(entry.timestamp))
-                            .font(.system(.caption, design: .monospaced))
-                            .foregroundColor(.secondary)
-                        Text(entry.kind.label)
-                            .font(.caption2.weight(.semibold))
-                            .foregroundColor(entry.kind.tintColor)
-                    }
-                    Text(entry.summary)
-                        .font(.callout)
-                        .lineLimit(isExpanded ? nil : 2)
-                }
-                Spacer(minLength: 0)
-                if canExpand {
-                    Image(systemName: "chevron.right")
-                        .rotationEffect(.degrees(isExpanded ? 90 : 0))
-                        .foregroundColor(.secondary)
-                        .font(.caption)
-                }
-            }
-            
-            if isExpanded, let detail = entry.detail {
-                Text(detail)
-                    .font(.system(.caption, design: .monospaced))
-                    .foregroundColor(.primary)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .padding(8)
-                    .background(Color(.systemGray6))
-                    .cornerRadius(6)
-            }
-        }
-        .contentShape(Rectangle())
-        .onTapGesture {
-            guard canExpand else { return }
-            withAnimation {
-                if isExpanded {
-                    expandedIds.remove(entry.id)
-                } else {
-                    expandedIds.insert(entry.id)
-                }
-            }
-        }
-    }
-    
-    // MARK: - Filtering
-    
+
+    // MARK: - Helpers
+
     private var filtered: [ConsoleEntry] {
-        // Newest first; the underlying array is appended chronologically.
         logStream.entries.reversed().filter { filter.includes($0.kind) }
     }
-    
-    // MARK: - Formatting
-    
-    private static let timeFormatter: DateFormatter = {
-        let f = DateFormatter()
-        f.dateFormat = "HH:mm:ss"
-        return f
-    }()
-    
-    private func timeString(_ date: Date) -> String {
-        Self.timeFormatter.string(from: date)
+
+    private func expansionBinding(for entry: ConsoleEntry) -> Binding<Bool> {
+        Binding(
+            get: { expandedIds.contains(entry.id) },
+            set: { newValue in
+                if newValue {
+                    expandedIds.insert(entry.id)
+                } else {
+                    expandedIds.remove(entry.id)
+                }
+            }
+        )
+    }
+
+    private func copyToPasteboard() {
+        UIPasteboard.general.string = ConsoleEntry.formatForExport(filtered)
+        withAnimation {
+            didJustCopy = true
+        }
+        DispatchQueue.main.asyncAfter(deadline: .now() + 1.5) {
+            withAnimation {
+                didJustCopy = false
+            }
+        }
     }
 }
