+++ Example/Example/Services/TripBuilderStore.swift
@@ -355,53 +355,3 @@ final class TripBuilderStore: ObservableObject {
     }
 }
 
-// MARK: - TripDestination
-
-enum TripDestination: Identifiable, Equatable {
-    case geofence(
-        id: String,
-        tag: String?,
-        externalId: String?,
-        displayName: String,
-        coord: CLLocationCoordinate2D
-    )
-    case coordinate(CLLocationCoordinate2D, label: String?)
-
-    var id: String {
-        switch self {
-        case .geofence(let id, _, _, _, _):
-            return "geofence:\(id)"
-        case .coordinate(let coord, _):
-            return String(format: "coord:%.6f:%.6f", coord.latitude, coord.longitude)
-        }
-    }
-
-    var displayName: String {
-        switch self {
-        case .geofence(_, _, _, let name, _):
-            return name
-        case .coordinate(let coord, let label):
-            return label ?? String(format: "%.4f, %.4f", coord.latitude, coord.longitude)
-        }
-    }
-
-    var coordinate: CLLocationCoordinate2D {
-        switch self {
-        case .geofence(_, _, _, _, let coord): return coord
-        case .coordinate(let coord, _): return coord
-        }
-    }
-
-    static func == (lhs: TripDestination, rhs: TripDestination) -> Bool {
-        lhs.id == rhs.id
-    }
-}
-
-// MARK: - TripEventMarker
-
-struct TripEventMarker: Identifiable {
-    let id: String
-    let coordinate: CLLocationCoordinate2D
-    let type: RadarEventType
-    let timestamp: Date
-}
